voice_keyboard_project/requirements.txt
torch
openai-whisper
pyaudio
keyboard
pynput
numpy
PyQt6
sounddevice